+++ push-spring-boot-starter/src/main/java/io/github/kk01001/push/MessagePushConfiguration.java
@@ -1,0 +1,13 @@
+package io.github.kk01001.push;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author linshiqiang
+ * @date 2024-12-25 14:43:00
+ * @description
+ */
+@Configuration(proxyBeanMethods = false)
+public class MessagePushConfiguration {
+
+}